[Refactor] #44 - View -> ViewController
